ipc: increase read timeout, flush proper buf number
The other side might want to do a PPS sync on startup which is cutting
it close with one second of read timeout.
Additionally the number of buffers read to ensure proper flushing was
previously not updated with the actual number of buffers, which led to
stale buffers with wrong timestamps being stuck in the ipc interface.

Change-Id: I2ec8d422ef31f45e97d9091e5d814429b101ffe8